sample_storage: move tests into a tests/ package, split by concern
Relocate backend_tests.py into tests/, matching the repo convention of a tests/
package with *_tests.py modules, and split the two suites (which share no
fixtures) into self-contained files:

- tests/driver_tests.py   - transport, queries, temperature/humidity (FakeSocket)
- tests/recovery_tests.py - pick/place fault classification + recovery (ScriptedSocket)
